+++ components/secondary/Events/RoundTableTablist.tsx
@@ -1,5 +1,4 @@
 import { FC } from "react";
-import { RoundtableCardProps } from "./RoundtableCard";
 
 interface Tab {
   label: string;
@@ -14,7 +13,6 @@ interface TabsListProps {
 
 export const TabsList: FC<TabsListProps> = ({ tabs, onTabClick }) => {
   return (
-
     <div
       role="tablist"
       aria-orientation="horizontal"
@@ -44,7 +42,6 @@ export const TabsList: FC<TabsListProps> = ({ tabs, onTabClick }) => {
           {tab.label}
         </button>
       ))}
-
     </div>
   );
 };
