+++ components/secondary/Events/RoundtableCard.tsx
@@ -2,11 +2,9 @@ import ClockIcon from "@/public/svg/clock_icon.svg"
 import LocationIcon from "@/public/svg/location_icon.svg"
 import CalendarIcon from "@/public/svg/calendar_icon.svg"
 import Link from "next/link";
-import { useMemo } from "react";
 
-export interface RoundtableCardProps {
-  startDate: string;
-  endDate?: string;
+interface RoundtableCardProps {
+  date: string;
   type: string;
   time: string;
   title: string;
@@ -21,7 +19,7 @@ export interface RoundtableCardProps {
 const RoundtableCard: React.FC<RoundtableCardProps> = ({
   imageUrl,
   label,
-  startDate,
+  date,
   type,
   title,
   description,
@@ -30,10 +28,6 @@ const RoundtableCard: React.FC<RoundtableCardProps> = ({
   buttonText = "Register to Attend",
   destinationUrl,
 }) => {
-
-  const formattedStartDate = useMemo(()=>{
-    return formatDate(startDate);
-  },[]) 
   return (
     <div className="grid md:grid-cols-3 h-full rounded-lg shadow-lg bg-white">
       <div className="relative h-full min-h-[200px] md:min-h-full">
@@ -46,7 +40,7 @@ const RoundtableCard: React.FC<RoundtableCardProps> = ({
             <div className="bg-affc-pink/90 text-white px-3 py-1 rounded inline-block mb-2">
               {label}
             </div>
-            <div className="text-lg font-medium font-inter">{formattedStartDate}</div>
+            <div className="text-lg font-medium font-inter">{date}</div>
             <div className="text-sm opacity-90 font-inter">{type}</div>
           </div>
         </div>
@@ -65,7 +59,7 @@ const RoundtableCard: React.FC<RoundtableCardProps> = ({
           </div>
           <div className="flex items-center gap-2">
             <CalendarIcon />
-            <span className="text-sm">{formattedStartDate}</span>
+            <span className="text-sm">{date}</span>
           </div>
           <div className="flex items-center gap-2">
             <LocationIcon className="text-affc-blue"/>
@@ -83,13 +77,5 @@ const RoundtableCard: React.FC<RoundtableCardProps> = ({
     </div>
   );
 };
-
-const formatDate = (dateString: string) => {
-  const date = new Date();
-
-  const options = { year: 'numeric', month: 'long', day: 'numeric' };
-  // @ts-expect-error is correct
-  return date.toLocaleDateString('en-US', options);
-}
 
 export default RoundtableCard
