+++ components/secondary/Events/index.ts
@@ -1,4 +1,3 @@
 export { default as RoundtableCard } from "./RoundtableCard"
-export { default as RoundTableEvents } from "./RoundTableEvents"
 export { default as SuggestRoundTable } from "./SuggestRoundTable"
 export { default as RoundTableTablist } from "./RoundTableTablist"
